Refactor: Move user management into own folder
